lumi as float

+++ totw.py
@@ -182,7 +182,7 @@ if __name__ == "__main__":
 
   parser.add_argument('--config', required=True, type=str, dest='config_file', metavar='<file.yml>', help='YAML file specifying input files and asssociated names')
   parser.add_argument('--weights', required=False, type=str, dest='weights_file', metavar='<file.json>', help='json file specifying the weights by dataset id')
-  parser.add_argument('--lumi', required=False, type=int, dest='global_luminosity', metavar='<ifb>', help='luminosity to use for scaling')
+  parser.add_argument('--lumi', required=False, type=float, dest='global_luminosity', metavar='<ifb>', help='luminosity to use for scaling')
 
   parser.add_argument('-i', '--input', dest='topLevel', type=str, help='Top level directory containing plots.', default='all')
   parser.add_argument('--file-ext', type=str, nargs='+', default=['root', 'pdf'], help='Output file extensions to make')
